+++ ft_printf/Source/pf_print_conv_o.c
@@ -29,7 +29,7 @@ t_pf_ret	pf_print_pad_conv_o(const char *str, t_pf_part *part ,t_pf_obj *obj)
 
 	len = ft_strlen(str);
 
-	num_len = (part->prcs > 0 && part->prcs > len) ? part->prcs : len;
+	num_len = (part->prcs > 0 && part->prcs > (int)len) ? part->prcs : len;
 	padding = (part->width >= 0) ? part->width : 0;
 	padding -= num_len;
 	padding = (padding < 0) ? 0 : padding;
