+++ ft_printf/Source/pf_print_num_full.c
@@ -16,7 +16,7 @@ t_pf_ret	print_num_full(const char *str, int num_len, t_pf_obj *obj)
 
 	len = ft_strlen(str);
 
-	if (num_len > len)
+	if (num_len > (int)len)
 		pf_print_nchar('0', num_len - len, obj);
 	obj->print(str, LEN_NS);
 	return (PF_RET_SUCCESS);
